+++ realm/realm-library/src/main/java/io/realm/RealmSchema.java
@@ -17,12 +17,14 @@
 package io.realm;
 
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.Map;
 import java.util.Set;
 
 import io.realm.internal.ColumnIndices;
 import io.realm.internal.ColumnInfo;
+import io.realm.internal.EmptyTableView;
 import io.realm.internal.ImplicitTransaction;
 import io.realm.internal.Table;
 import io.realm.internal.Util;
@@ -121,6 +123,30 @@ public final class RealmSchema {
     }
 
     /**
+     * When a Table is removed or renamed, its corresponding cache entries should also be removed
+     * properly to ensure derived {@link RealmResults} or {@link RealmList} to have correct values.
+     *
+     * @param table a Table to be removed from cache
+     */
+    private void removeTableFromTableCache(final Table table) {
+        // firstly remove the entry from dynamic cache
+        for (Iterator<Map.Entry<String, Table>> itr = dynamicClassToTable.entrySet().iterator(); itr.hasNext();) {
+            Map.Entry<String, Table> entry = itr.next();
+            if (entry.getValue().equals(table)) {
+                itr.remove();
+            }
+        }
+
+        // then remove the entry from typed cache
+        for (Iterator<Map.Entry<Class<? extends RealmModel>, Table>> itr = classToTable.entrySet().iterator(); itr.hasNext();) {
+            Map.Entry<Class<? extends RealmModel>, Table> entry = itr.next();
+            if (entry.getValue().equals(table)) {
+                itr.remove();
+            }
+        }
+    }
+
+    /**
      * Removes a class from the Realm. All data will be removed. Removing a class while other classes point
      * to it will throw an {@link IllegalStateException}. Remove those classes or fields first.
      *
@@ -131,6 +157,12 @@ public final class RealmSchema {
         String internalTableName = TABLE_PREFIX + className;
         checkHasTable(className, "Cannot remove class because it is not in this Realm: " + className);
         Table table = getTable(className);
+        // create a final, immutable empty TableView
+        final EmptyTableView emptyTableView = new EmptyTableView(table);
+        // invalidate all the RealmResults related to this class
+        RealmCache.invalidateRemovedClassFromCachedRealm(realm.getConfiguration(), emptyTableView);
+        // remove corresponding table from cache if exists
+        removeTableFromTableCache(table);
         if (table.hasPrimaryKey()) {
             table.setPrimaryKey(null);
         }